rename: 컴포넌트 폴더명 변경
- 파스칼 케이스로 통일
