+++ src/components/common/pagination/PaginationButton.tsx
@@ -3,7 +3,7 @@ import { type VariantProps } from 'class-variance-authority';
 import clsx from 'clsx';
 import { FC, ReactNode } from 'react';
 
-import { paginationButtonVariants } from '@/components/common/pagination/PaginationButtonVariants';
+import { paginationButtonVariants } from '@/components/common/Pagination/PaginationButtonVariants';
 
 interface PaginationButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement>,
