+++ src/components/index.ts
@@ -1,25 +1,25 @@
-export { default as Button } from '@/components/common/button/Button';
-export { default as Divider } from '@/components/common/divider/Divider';
+export { default as Button } from '@/components/common/Button/Button';
+export { default as Divider } from '@/components/common/Divider/Divider';
 export { default as DropdownMenu } from '@/components/common/Dropdown/DropdownMenu';
 export { default as DropdownMenuList } from '@/components/common/Dropdown/DropdownMenuList';
-export { default as Icon } from '@/components/common/icon/Icon';
-export { default as Modal } from '@/components/common/modal/Modal';
-export { default as PageHeader } from '@/components/common/pageheader/PageHeader';
-export { default as PageHeaderLeft } from '@/components/common/pageheader/PageHeaderLeft';
-export { default as PageHeaderLegacy } from '@/components/common/pageheader/PageHeaderLegacy';
-export { default as PageHeaderRight } from '@/components/common/pageheader/PageHeaderRight';
-export { default as Pagination } from '@/components/common/pagination/Pagination';
-export { default as PaginationButton } from '@/components/common/pagination/PaginationButton';
-export { default as Spacer } from '@/components/common/spacer/Spacer';
+export { default as Icon } from '@/components/common/Icon/Icon';
+export { default as Modal } from '@/components/common/Modal/Modal';
+export { default as PageHeader } from '@/components/common/pageHeader/PageHeader';
+export { default as PageHeaderLeft } from '@/components/common/pageHeader/PageHeaderLeft';
+export { default as PageHeaderLegacy } from '@/components/common/pageHeader/PageHeaderLegacy';
+export { default as PageHeaderRight } from '@/components/common/pageHeader/PageHeaderRight';
+export { default as Pagination } from '@/components/common/Pagination/Pagination';
+export { default as PaginationButton } from '@/components/common/Pagination/PaginationButton';
+export { default as Spacer } from '@/components/common/Spacer/Spacer';
 export { default as TextField } from '@/components/common/text/TextField';
 export { default as Textarea } from '@/components/common/text/Textarea';
-export { default as TextBox } from '@/components/common/textbox/TextBox';
-export { default as Toast } from '@/components/common/toast/Toast';
+export { default as TextBox } from '@/components/common/Textbox/TextBox';
+export { default as Toast } from '@/components/common/Toast/Toast';
 
 // Blog Components
-export { default as PostBody } from '@/components/blog/post/PostBody';
-export { default as PostCard } from '@/components/blog/post/PostCard';
-export { default as PostHeader } from '@/components/blog/post/PostHeader';
+export { default as PostBody } from '@/components/blog/Post/PostBody';
+export { default as PostCard } from '@/components/blog/Post/PostCard';
+export { default as PostHeader } from '@/components/blog/Post/PostHeader';
 
 // UI Components
 export { default as Menu } from '@/components/ui/Menu';
