+++ src/playground/Playground.tsx
@@ -1,18 +1,18 @@
 import { CreateIcon } from '@/assets/icons/common';
-import PostBody from '@/components/blog/post/PostBody';
-import PostCard from '@/components/blog/post/PostCard';
-import PostHeader from '@/components/blog/post/PostHeader';
+import PostBody from '@/components/blog/Post/PostBody';
+import PostCard from '@/components/blog/Post/PostCard';
+import PostHeader from '@/components/blog/Post/PostHeader';
 import DropdownMenu from '@/components/common/Dropdown/DropdownMenu';
-import Button from '@/components/common/button/Button';
-import Divider from '@/components/common/divider/Divider';
-import Modal from '@/components/common/modal/Modal';
-import PageHeader from '@/components/common/pageheader/PageHeader';
-import PageHeaderLegacy from '@/components/common/pageheader/PageHeaderLegacy';
-import Pagination from '@/components/common/pagination/Pagination';
+import Button from '@/components/common/Button/Button';
+import Divider from '@/components/common/Divider/Divider';
+import Modal from '@/components/common/Modal/Modal';
+import PageHeader from '@/components/common/pageHeader/PageHeader';
+import PageHeaderLegacy from '@/components/common/pageHeader/PageHeaderLegacy';
+import Pagination from '@/components/common/Pagination/Pagination';
 import TextField from '@/components/common/text/TextField';
 import Textarea from '@/components/common/text/Textarea';
-import TextBox from '@/components/common/textbox/TextBox';
-import Toast from '@/components/common/toast/Toast';
+import TextBox from '@/components/common/Textbox/TextBox';
+import Toast from '@/components/common/Toast/Toast';
 import Sidebar from '@/layout/Sidebar';
 import { useState } from 'react';
